dc testcase: Add no-op support for cloudflare-go server
This change refactors the cloudflare-go runner and adds server support.
It also provides rudimentary support for the "dc" test case; the server
is run, but it currently doesn't configure a DC.

This change also modifies the Dockerfile for cloudflare-go so that the
runner is built into the image, rather than interpreted on the fly. This
ensures that a client-side build error doesn't cause the server to hang.